Updated Map file structure to be more intuitive, added ReserveMap component to Reserve page
