+++ client/src/Map/Layers/TileLayer.js
@@ -1,5 +1,5 @@
 import {useContext, useEffect} from 'react'
-import MapContext from './MapContext'
+import MapContext from '../MapContext'
 import OLTileLayer from 'ol/layer/Tile'
