+++ client/src/Reserve/ReserveMap.js
@@ -1,0 +1,83 @@
+import React, { useState, useEffect } from 'react';
+import '../App.css';
+import Map from '../Map/Map.js';
+import Layers from '../Map/Layers/Layers.js';
+import TileLayer from '../Map/Layers/TileLayer.js';
+import VectorLayer from '../Map/Layers/VectorLayer.js';
+import { Circle as CircleStyle, Fill, Stroke, Style } from 'ol/style';
+import { osm, vector, xyz } from '../Map/Wrappers/wrapper.js';
+import { fromLonLat, get } from 'ol/proj';
+import GeoJSON from 'ol/format/GeoJSON';
+import Controls from '../Map/Controls/Controls.js';
+import FullScreenControl from '../Map/Controls/FullScreenControl';
+import KML from 'ol/format/KML.js';
+import VectorSource from 'ol/source/Vector.js';
+import KMLVectorLayer from '../Map/Layers/KMLVectorLayer.js';
+
+let styles = {
+  MultiPolygon: new Style({
+    stroke: new Stroke({
+      color: 'blue',
+      width: 1,
+    }),
+    fill: new Fill({
+      color: 'rgba(0,0,255,0.1)',
+    }),
+  }),
+};
+
+const samGeoObject = {
+	type: 'Feature',
+	geometry: {
+		'type': 'Point',
+		'coordinates': [-146.44166473513687, 64.31714411488758]
+	},
+	'properties': {
+		'name': 'Test Object'
+	}
+
+	}
+const samGeoObject2 = {
+	type: 'Feature',
+	geometry: {
+		'type': 'Point',
+		'coordinates': [-147.44166473513687, 65.31714411488758]
+	},
+	'properties': {
+		'name': 'Test Object'
+	}
+
+	}
+
+
+const ReserveMap = () => {
+
+  const [center, setCenter] = useState([
+    -146.44166473513687, 64.31714411488758,
+  ]);
+  const [zoom, setZoom] = useState(8);
+
+  return (
+    <div>
+      <Map center={fromLonLat(center)} zoom={zoom}>
+        <Layers>
+          <TileLayer
+            source={xyz({
+              url: 'https://services.arcgisonline.com/arcgis/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}',
+            })}
+            zIndex={0}
+          />
+					<VectorLayer
+							source={vector({ features: new GeoJSON().readFeatures(samGeoObject, {featureProjection: get('EPSG:3857') }) })}
+							/>
+
+          <KMLVectorLayer zIndex={99} />
+        </Layers>
+        <Controls>
+          <FullScreenControl />
+        </Controls>
+      </Map>
+    </div>
+  );
+};
+export default ReserveMap;
